Important: Changes knownlist updating from move based to time based. Find and forget are run every 1.5 seconds but not at the same time (also forget could be later changed to run even less often, like every 3 seconds). The old move based update had several little bugs: - Static l2Character gathered bigger and bigger knownlist unless an update was run. A temporary update was run with social animations code but it wasn't a nice way. Either a separate forget task would've been needed OR characters, when they move, also should run a forget task to all objects in nearby world regions (might still not help sometimes). This fix would've increased cpu use. - Some unneeded temporary lists, null checks and several isVisible checks - Mobs forgot each other (start area spawns) when moving - Guards, if were set to attack aggro mobs (retail: false), didn't see moving mobs. - Near-exploitable delays in knownlist update when moving fast. Partly because neighbor world regions were turned on after 30 sec wait! I changed that to 1 sec. Also Mob AI wasn't running there...
Old and current still don't update mob knownlist when mob moves far away. This could be added but
by far nobody has seemed to need it. Factions, minions etc are always in their list.

Move-based knownlist update was very cpu effective in certain situations (when nobody moves) but 
with the required fixes (the list above) it would've come down to about this time-based.
When chars move a lot, time-based is more effective. The old code could also be optimized with
the fixes in this patch (revert not needed, just a couple of old lines back) if CPU use noticably rises.

This patch reduces memory use after startup around 35 MB.