Fix for clan pledge, by nitrozik:
''There is a problem with level 10 clan. Instead of increasing the rank of players it doesn't recognize their rank. For example a clan leader of a lv 8 clan is a Marquis with full clan skills, after making it level 10 hes a Baron with limited clan skills. All the royal guard and knight captains are also baron's with limited clan skills. The clan has full skills and atm only 10 of them are available due to everyone in the clan being a baron including heroes, captains and clan leader.''